Several errors corrected. New event processing for Rosenbrock solver.

+++ trunk/src/org/simulator/sbml/astnode/ASTNodeInterpreterWithTime.java
@@ -458,7 +458,7 @@ public class ASTNodeInterpreterWithTime {
   }
   
   public double frac(int numerator, int denominator) {
-    return (numerator / denominator);
+    return ((double)numerator / (double)denominator);
   }
   
   public double frac(ASTNodeObject left, ASTNodeObject right, double time)
